+++ core/datastore/src/main/java/com/threegap/bitnagil/datastore/auth/serializer/AuthTokenSerializer.kt
@@ -1,0 +1,6 @@
+package com.threegap.bitnagil.datastore.auth.serializer
+
+import androidx.datastore.core.Serializer
+import com.threegap.bitnagil.datastore.auth.model.AuthToken
+
+interface AuthTokenSerializer : Serializer<AuthToken>
