--- core/datastore/src/main/java/com/threegap/bitnagil/datastore/serializer/TokenSerializer.kt
@@ -1,6 +1,0 @@
-package com.threegap.bitnagil.datastore.serializer
-
-import androidx.datastore.core.Serializer
-import com.threegap.bitnagil.datastore.model.AuthToken
-
-interface TokenSerializer : Serializer<AuthToken>
